riscv(ffi): FFI init
Largely copied from LoongArch because RV LP64D and LA LP64D are similar, with
the main difference being float NaN boxing.

With this patch, LuaJIT in the interpreter mode can now pass OpenResty FFI test.

Despite the existence of a few callback-related changes, JIT and callback are
still in the WIP stage, so do not anticipate that part to function at this time.

This is a WIP, archive only.