loong: prepare for RISC-V cross check
Regression test kit result on LoongArchLinux + QEMU Virt platform suggests
that this LA port is a functional New-World ABI port.
For the RISC-V port, debugging using the LA port as a known good reference
is a simple but effective solution because the RISC-V port refers to it.

We intend to change this for better interpreter+FFI cross check experience.
DynASM Instruction mnemonics, VM Instructions, DISPATCH to GL migration, and
slt+b fusion are the anticipated changes.