riscv(emit): emit_loadu64 regression workaround
The approach here is a greedy algorithm. In the worst case, we might get
7-instrs instead of 6-instrs of the naive approach, a regression!
Therefore fallback to naive approach if the algorithm can't do better.

This might be a plug-dp problem, but I'm exhausted. Take this as-is for now.
Also I don't believe there would be any worse case with 5 or less steps,
plus this commit already provides significant performance boost anyway.

This is a WIP, archive only.